pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rootcauseai"]

[project]
name = "rootcauseai"
version = "0.1.0"
description = "RootCauseAI - Log Analyzer Agent. AI-powered log analysis tool built with LangChain and OpenAI"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "RootCauseAI Contributors"}
]
keywords = ["root-cause", "log-analysis", "ai", "langchain", "openai", "fastapi", "rootcauseai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langchain>=0.3.7",
    "langchain-core>=0.3.21",
    "langchain-community>=0.3.7",
    "langchain-openai>=0.2.10",
    "langchain-text-splitters>=0.3.2",
    "openai>=1.57.0",
    "tiktoken>=0.8.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",  # Required for FastAPI TestClient
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
rootcauseai = "rootcauseai.app:main"

[project.urls]
Homepage = "https://github.com/yourusername/rootcauseai"
Documentation = "https://github.com/yourusername/rootcauseai#readme"
Repository = "https://github.com/yourusername/rootcauseai"
Issues = "https://github.com/yourusername/rootcauseai/issues"

# Ruff configuration (replaces Black, isort, flake8, etc.)
[tool.ruff]
target-version = "py311"  # Supports Python 3.11-3.13
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate (commented code)
    "PD",  # pandas-vet
    "PL",  # Pylint
    "TRY", # tryceratops
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "PLR0913", # too many arguments (allow for FastAPI routes)
    "PLR2004", # magic value used in comparison
    "TRY003", # avoid specifying long messages outside exception class
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["rootcauseai"]

[tool.ruff.mccabe]
max-complexity = 10

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_core.*",
    "langchain_community.*",
    "langchain_openai.*",
    "langchain_text_splitters.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=rootcauseai",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["rootcauseai"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
